Added file middleware
Migrated configuration of MIME_TYPE_MAP and multer to the file middleware.
